adding abstract class for listeners
